removed outdated PeriodicTable object
