--- data/elements.json
@@ -1,195 +1,0 @@
-{
-    "H": {},
-    "He": {},
-    "Li": {},
-    "Be": {},
-    "B": {},
-    "C": {
-        "atomic_number": 6,
-        "atomic_radius": 67,
-        "atomic_weight": 12.001,
-        "covalent_radius": 76,
-        "lattice_constants": {
-            "hcp": [246.4, 246.4, 671.1]
-        }
-    },
-    "N": {},
-    "O": {
-        "atomic_number": 8,
-        "atomic_radius": 48,
-        "atomic_weight": 15.999,
-        "covalent_radius": 66
-    },
-    "F": {},
-    "Ne": {},
-    "Na": {},
-    "Mg": {
-        "atomic_number": 12,
-        "atomic_radius": 145,
-        "atomic_weight": 24.305,
-        "covalent_radius": 141,
-        "lattice_constants": {
-            "hcp": [320.94, 302.94, 521.08]
-        }
-    },
-    "Al": {
-        "atomic_number": 13,
-        "atomic_radius": 118,
-        "atomic_weight": 26.982,
-        "covalent_radius": 121,
-        "lattice_constants": {
-            "fcc": [404.95, 404.95, 404.95]
-        }
-    },
-    "Si": {},
-    "P": {},
-    "S": {},
-    "Cl": {},
-    "Ar": {},
-    "K": {},
-    "Ca": {},
-    "Sc": {},
-    "Ti": {
-        "atomic_number": 22,
-        "atomic_radius": 176,
-        "atomic_weight": 47.867,
-        "covalent_radius": 160,
-        "lattice_constants": {
-            "hcp": [295.08, 295.08, 468.55]
-        }
-    },
-    "V": {
-        "atomic_number": 23,
-        "atomic_radius": 171,
-        "atomic_weight": 50.942,
-        "covalent_radius": 153,
-        "lattice_constants": {
-            "bcc": [303.0, 303.0, 303.0]
-        }
-    },
-    "Cr": {
-        "atomic_number": 24,
-        "atomic_radius": 166,
-        "atomic_weight": 51.996,
-        "covalent_radius": 139,
-        "lattice_constants": {
-            "bcc": [291.0, 291.0, 291.0]
-        }
-    },
-    "Mn": {
-        "atomic_number": 25,
-        "atomic_radius": 161,
-        "atomic_weight": 54.938,
-        "covalent_radius": 139,
-        "lattice_constants": {
-            "bcc": [891.25, 891.25, 891.25]
-        }
-    },
-    "Fe": {
-        "atomic_number": 26,
-        "atomic_radius": 156,
-        "atomic_weight": 55.845,
-        "covalent_radius": 132,
-        "lattice_constants": {
-            "bcc": [286.65, 286.65, 286.65]
-        }
-    },
-    "Co": {
-        "atomic_number": 27,
-        "atomic_radius": 152,
-        "atomic_weight": 58.933,
-        "covalent_radius": 126,
-        "lattice_constants": {
-            "hcp": [250.71, 250.71, 406.95]
-        }
-    },
-    "Ni": {
-        "atomic_number": 28,
-        "atomic_radius": 149,
-        "atomic_weight": 58.693,
-        "covalent_radius": 124,
-        "lattice_constants": {
-            "fcc": [352.4, 352.4, 352.4]
-        }
-    },
-    "Cu": {},
-    "Zn": {},
-    "Ga": {},
-    "Ge": {},
-    "As": {},
-    "Se": {},
-    "Br": {},
-    "Kr": {},
-    "Rb": {},
-    "Sr": {},
-    "Y": {},
-    "Zr": {},
-    "Nb": {},
-    "Mo": {
-        "atomic_number": 42,
-        "atomic_radius": 190,
-        "atomic_weight": 95.950,
-        "covalent_radius": 154,
-        "lattice_constants": {
-            "bcc": [314.7, 314.7, 314.7]
-        }
-    },
-    "Tc": {},
-    "Ru": {},
-    "Rh": {},
-    "Pd": {
-        "atomic_number": 46,
-        "atomic_radius": 169,
-        "atomic_weight": 106.420,
-        "covalent_radius": 139,
-        "lattice_constants": {
-            "fcc": [389.07, 389.07, 389.07]
-        }
-    },
-    "Ag": {},
-    "Cd": {},
-    "In": {},
-    "Sn": {},
-    "Sb": {},
-    "Te": {},
-    "I": {},
-    "Xe": {},
-    "Cs": {},
-    "Ba": {},
-    "La": {},
-    "Ce": {},
-    "Pr": {},
-    "Nd": {},
-    "Pm": {},
-    "Sm": {},
-    "Eu": {},
-    "Gd": {},
-    "Tb": {},
-    "Dy": {},
-    "Ho": {},
-    "Er": {},
-    "Tm": {},
-    "Yb": {},
-    "Lu": {},
-    "Hf": {},
-    "Ta": {},
-    "W": {},
-    "Re": {},
-    "Os": {},
-    "Ir": {},
-    "Pt": {},
-    "Au": {},
-    "Hg": {},
-    "Tl": {},
-    "Pb": {},
-    "Bi": {},
-    "Po": {},
-    "At": {},
-    "Rn": {},
-    "Fr": {},
-    "Ra": {},
-    "Ac": {},
-    "Th": {},
-    "Pa": {},
-    "U": {}
-}
